Added bluetooth service (broken)
