+++ app/src/main/java/ca/sfu/BlueRadar/ui/devices/DevicesFragment.kt
@@ -21,9 +21,11 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import ca.sfu.BlueRadar.databinding.FragmentDevicesBinding
+import ca.sfu.BlueRadar.services.BluetoothService
 import ca.sfu.BlueRadar.services.DatabaseService
 import ca.sfu.BlueRadar.services.LocationTrackingService
 import ca.sfu.BlueRadar.ui.devices.data.Device
+import ca.sfu.BlueRadar.util.Util
 import com.google.android.gms.maps.model.LatLng
 import nl.bryanderidder.themedtogglebuttongroup.ThemedToggleButtonGroup
 
@@ -31,9 +33,6 @@ class DevicesFragment : Fragment() {
 
     //Global Variables
     private var _binding: FragmentDevicesBinding? = null
-    private var deviceNameList: ArrayList<String> = ArrayList()
-    private var bluetoothDevices: ArrayList<String> = ArrayList()
-    private var deviceAddresses: ArrayList<String> = ArrayList()
     private var viewDevices = 0
     private val binding get() = _binding!!
 
@@ -47,113 +46,6 @@ class DevicesFragment : Fragment() {
     private lateinit var recyclerAdapter: DeviceRecyclerAdapter
     private lateinit var buttonGroup: ThemedToggleButtonGroup
 
-    //Bluetooth broadcast receiver
-    private val receiver = object : BroadcastReceiver() {
-
-        override fun onReceive(context: Context, intent: Intent) {
-            when (intent.action) {
-                BluetoothDevice.ACTION_ACL_CONNECTED -> {
-
-                    //Check if any paired bluetooth devices are available
-                    val device: BluetoothDevice? = intent.getParcelableExtra(
-                        BluetoothDevice
-                            .EXTRA_DEVICE
-                    )
-                    val temp = deviceViewModel.allEntriesLiveData.value
-                    var currentLoc = LatLng(0.0, 0.0)
-                    LocationTrackingService.currentPoint.observe(viewLifecycleOwner) {
-                        currentLoc = it
-                    }
-
-                    //Save device to database on connect if not already available (new device sync)
-                    if (temp?.isNotEmpty() == true && device != null) {
-                        for (i in temp) {
-                            if (i.deviceName == device.name) {
-
-                                i.deviceConnected = true
-                                deviceViewModel.update(i)
-                                if (i.deviceTracking == true) {
-                                    i.deviceLastLocation = currentLoc
-                                }
-                                updateRecyclerView()
-                                val toast: Toast = Toast.makeText(
-                                    requireContext(),
-                                    "Current Loc: ${currentLoc.latitude}, ${currentLoc.longitude}",
-                                    Toast
-                                        .LENGTH_SHORT
-                                )
-                                toast.show()
-                            }
-                        }
-                    }
-                    Log.d("BluetoothReceiver", "BluetoothDevice ${device?.name} connected")
-                }
-
-                //Save latlng upon disconnect and update status
-                BluetoothDevice.ACTION_ACL_DISCONNECTED -> {
-                    val g: ArrayList<LatLng?> = ArrayList()
-                    g.add(LocationTrackingService.currentPoint.value)
-                    val device: BluetoothDevice? = intent.getParcelableExtra(
-                        BluetoothDevice
-                            .EXTRA_DEVICE
-                    )
-                    val dbList = deviceViewModel.allEntriesLiveData.value
-                    var lastLoc = LatLng(0.0, 0.0)
-                    LocationTrackingService.currentPoint.observe(viewLifecycleOwner) {
-                        lastLoc = it
-                    }
-                    if (dbList?.isNotEmpty() == true && device != null) {
-                        for (i in dbList) {
-                            if (i.deviceName == device.name) {
-                                i.deviceConnected = false
-                                i.deviceLastLocation = lastLoc
-                                deviceViewModel.update(i)
-                                updateRecyclerView()
-                                val toast: Toast = Toast.makeText(
-                                    requireContext(),
-                                    "Last Loc: ${lastLoc.latitude}, ${lastLoc.longitude}",
-                                    Toast
-                                        .LENGTH_SHORT
-                                )
-                                toast.show()
-                            }
-                        }
-                    }
-                    for (i in deviceViewModel.allEntriesLiveData.value!!) {
-                        Log.d("check me", i.toString())
-                    }
-                    Log.d("BluetoothReceiver", "BluetoothDevice ${device?.name} disconnected")
-
-                }
-
-                //Add device to arrayList
-                BluetoothDevice.ACTION_FOUND -> {
-                    // Discovery has found a device. Get the BluetoothDevice
-                    // object and its info from the Intent.
-                    val device: BluetoothDevice? =
-                        intent.getParcelableExtra(BluetoothDevice.EXTRA_DEVICE)
-                    val deviceName = device?.name
-                    val deviceHardwareAddress = device?.address // MAC address
-                    val rssi: String = intent.getShortExtra(
-                        BluetoothDevice.EXTRA_RSSI, Short
-                            .MIN_VALUE
-                    ).toString()
-                    if (!deviceAddresses.contains(deviceHardwareAddress)) {
-                        if (deviceHardwareAddress != null) {
-                            deviceAddresses.add(deviceHardwareAddress)
-                        }
-                        val deviceString = if (deviceName.isNullOrEmpty()) {
-                            "$deviceHardwareAddress RSSI $rssi dBm"
-                        } else {
-                            "$deviceName RSSI $rssi dBm"
-                        }
-                        bluetoothDevices.add(deviceString)
-                    }
-                }
-            }
-        }
-    }
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         bluetoothManager = getSystemService(requireContext(), BluetoothManager::class.java)!!
@@ -166,60 +58,17 @@ class DevicesFragment : Fragment() {
             )
         }
 
-        // Register for broadcasts when a device is discovered.
-        val filter = IntentFilter().apply {
-            addAction(BluetoothAdapter.ACTION_DISCOVERY_STARTED)
-            addAction(BluetoothAdapter.ACTION_DISCOVERY_FINISHED)
-            addAction(BluetoothDevice.ACTION_FOUND)
-            addAction(BluetoothDevice.ACTION_ACL_CONNECTED)
-            addAction(BluetoothDevice.ACTION_ACL_DISCONNECTED)
-            addAction(BluetoothAdapter.ACTION_STATE_CHANGED)
-            addAction(BluetoothAdapter.ACTION_CONNECTION_STATE_CHANGED)
-        }
-        requireActivity().registerReceiver(receiver, filter)
+        requireActivity().registerReceiver(BluetoothService.receiver, Util.filter)
         firstBootSetup()
     }
 
     private fun firstBootSetup() {
-        deviceViewModel =
-            ViewModelProvider(requireActivity(), DatabaseService.viewModelFactory)[DeviceViewModel::class.java]
+        BluetoothService.deviceViewModel = ViewModelProvider(
+            requireActivity(),
+            DatabaseService.viewModelFactory
+        )[DeviceViewModel::class.java]
 
-        val pairedDevices: Set<BluetoothDevice>? = bluetoothAdapter.bondedDevices
-        pairedDevices?.forEach { device ->
-            val deviceName = device.name
-            deviceNameList.add(deviceName)
-            val btDevice = Device()
-            btDevice.deviceName = deviceName
-            btDevice.deviceType = device.type.toString()
-            btDevice.deviceMacAddress = device.address
-
-            var liveList = deviceViewModel.allEntriesLiveData.value
-            var isDuplicate = false
-
-            if (liveList != null) {
-                for (check in liveList) {
-                    if (check.deviceName == btDevice.deviceName)
-                        isDuplicate = true
-                }
-            }
-
-            if (!isDuplicate) {
-                deviceViewModel.insert(btDevice)
-            }
-
-            val deviceHardwareAddress = device.address // MAC Address
-            Log.d("bonded-device-name", btDevice.deviceName)
-            Log.d("bonded-device-address", deviceHardwareAddress)
-        }
-
-        // checking status
-        val toast: Toast
-        if (!bluetoothAdapter.isEnabled) {
-            Log.d("bluetooth-checker", "Bluetooth is Disabled")
-            toast = Toast.makeText(requireContext(), "Bluetooth is Disabled", Toast.LENGTH_SHORT)
-            toast.show()
-        }
-        bluetoothAdapter.startDiscovery()
+        deviceViewModel = BluetoothService.deviceViewModel
     }
 
     private fun setupButtonGroupListener() {
@@ -335,7 +184,7 @@ class DevicesFragment : Fragment() {
         super.onDestroy()
         _binding = null
         //deviceViewModel.deleteAll()
-        requireActivity().unregisterReceiver(receiver)
+        requireActivity().unregisterReceiver(BluetoothService.receiver)
     }
 
     class MarginItemDecoration(private val spaceSize: Int) : RecyclerView.ItemDecoration() {
